some spring-cleaning of tests.spark fixture now on conftest.py
