+++ tests/conftest.py
@@ -480,3 +480,54 @@ def sqlite_con_4(gamma_settings_4):
 
 
     yield con
+
+    
+    
+
+@pytest.fixture(scope="module")
+def spark():
+
+    try:
+        import pyspark
+        from pyspark import SparkContext, SparkConf
+        from pyspark.sql import SparkSession
+        from pyspark.sql import types
+
+        conf = SparkConf()
+
+        conf.set("spark.sql.shuffle.partitions", "1")
+        conf.set("spark.jars.ivy", "/home/jovyan/.ivy2/")
+        conf.set("spark.driver.extraClassPath", "jars/scala-udf-similarity-0.0.6.jar")
+        conf.set("spark.jars", "jars/scala-udf-similarity-0.0.6.jar")
+        conf.set("spark.driver.memory", "4g")
+        conf.set("spark.sql.shuffle.partitions", "24")
+
+        sc = SparkContext.getOrCreate(conf=conf)
+
+        spark = SparkSession(sc)
+
+        udfs = [
+            ("jaro_winkler_sim", "JaroWinklerSimilarity", types.DoubleType()),
+            ("jaccard_sim", "JaccardSimilarity", types.DoubleType()),
+            ("cosine_distance", "CosineDistance", types.DoubleType()),
+            ("Dmetaphone", "DoubleMetaphone", types.StringType()),
+            ("QgramTokeniser", "QgramTokeniser", types.StringType()),
+            ("Q3gramTokeniser", "Q3gramTokeniser", types.StringType()),
+            ("Q4gramTokeniser", "Q4gramTokeniser", types.StringType()),
+            ("Q5gramTokeniser", "Q5gramTokeniser", types.StringType()),
+        ]
+
+        for a, b, c in udfs:
+            spark.udf.registerJavaFunction(a, "uk.gov.moj.dash.linkage." + b, c)
+        SPARK_EXISTS = True
+    except:
+        SPARK_EXISTS = False
+
+    if SPARK_EXISTS:
+        print("Spark exists, running spark tests")
+        yield spark
+    else:
+        spark = None
+        logger.error("Spark not available")
+        print("Spark not available")
+        yield spark
